requirements.txt
psycopg2-binary
PyPika
python-dotenv